renamed mask module utilities to utils
