--- docs/source/eolearn.mask.utilities.rst
@@ -1,7 +1,0 @@
-eolearn.mask.utilities
-======================
-
-.. automodule:: eolearn.mask.utilities
-    :members:
-    :undoc-members:
-    :show-inheritance:
