+++ docs/source/eolearn.mask.utils.rst
@@ -1,0 +1,7 @@
+eolearn.mask.utils
+==================
+
+.. automodule:: eolearn.mask.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
