+++ mask/eolearn/mask/__init__.py
@@ -5,7 +5,7 @@ Public classes and functions of mask subpackage
 from .cloud_mask import CloudMaskTask
 from .masking import MaskFeatureTask, JoinMasksTask
 from .snow_mask import SnowMaskTask, TheiaSnowMaskTask
-from .utilities import resize_images
+from .utils import resize_images
 from .mask_counting import ClassFrequencyTask
 
 __version__ = "1.0.0"
