+++ mask/eolearn/mask/cloud_mask.py
@@ -22,7 +22,7 @@ from skimage.morphology import disk
 from sentinelhub import bbox_to_resolution
 from eolearn.core import EOTask, FeatureType, execute_with_mp_lock
 
-from .utilities import map_over_axis, resize_images
+from .utils import map_over_axis, resize_images
 
 LOGGER = logging.getLogger(__name__)
